agentic-workspace/requirements.txt
# Valtheron Agentic Workspace Dependencies

# LLM Provider - Anthropic
anthropic>=0.40.0

# Configuration
pyyaml>=6.0

# CLI
argparse

# Async support
asyncio

# Optional: Testing frameworks (for test_runner tool)
pytest>=7.0.0
pytest-json-report>=1.5.0
pytest-cov>=4.0.0

# Optional: Development tools
black>=23.0.0
mypy>=1.0.0
ruff>=0.1.0
